tests/wpt/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.measure.fontBoundingBox.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.text.measure.fontBoundingBox</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("/fonts/CanvasTest.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.fontBoundingBox</h1>
<p class="desc">Testing fontBoundingBox</p>


<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing fontBoundingBox");
_addTest(function(canvas, ctx) {

  deferTest();
  var f = new FontFace("CanvasTest", "/fonts/CanvasTest.ttf");
  document.fonts.add(f);
  document.fonts.ready.then(() => {
      step_timeout(t.step_func_done(function () {
          ctx.font = '50px CanvasTest';
          ctx.direction = 'ltr';
          ctx.align = 'left'
          // approx_equals because font metrics may be rounded slightly differently by different platforms/browsers.
          assert_approx_equals(ctx.measureText('A').fontBoundingBoxAscent, 50 * 1745 / 1024, 1, "unexpected fontBoundingBoxAscent");
          assert_approx_equals(ctx.measureText('A').fontBoundingBoxDescent, 50 * 805 / 1024, 1, "unexpected fontBoundingBoxDescent");

          assert_approx_equals(ctx.measureText('ABCD').fontBoundingBoxAscent, 50 * 1745 / 1024, 1, "unexpected fontBoundingBoxAscent");
          assert_approx_equals(ctx.measureText('ABCD').fontBoundingBoxDescent, 50 * 805 / 1024, 1, "unexpected fontBoundingBoxDescent");
      }), 500);
  });

});
</script>
