Safari-611.1.10/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.measure.fontBoundingBox-expected.txt

2d.text.measure.fontBoundingBox
Testing fontBoundingBox
Actual output:

FAIL Testing fontBoundingBox assert_equals: ctx.measureText('A').fontBoundingBoxAscent === 85 (got 86[number], expected 85[number]) expected 85 but got 86
